Update firmwares/ws2812b-to-wb
Обновлён код скетча ws2812b-to-wb, добавил конфиг проекта platformio.ini
