+++ firmwares/ws2812b-to-wb/README.md
@@ -2,6 +2,8 @@
 
 Инструкция по созданию устройства на контроллере ESP32 для управления лентой WS2812B и подключению этого устройства к контроллеру Wiren Board.
 
-Скетч для Arduino находится в файле `main.c`, а виртуальное устройство для контроллера Wiren Board в файле `ws2812b-wb-vidtualdevice.js`.
+Скетч для Arduino находится в файле `src/main.cpp`, а виртуальное устройство для контроллера Wiren Board в файле `ws2812b-wb-vidtualdevice.js`.
+
+Также прилагается файл с настройками проекта `platformio.ini`
 
 Инструкция по сборке устройства: https://wirenboard.com/wiki/Address_strip_WS2812B
