+++ firmwares/ws2812b-to-wb/platformio.ini
@@ -1,0 +1,19 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:esp32doit-devkit-v1]
+platform = espressif32
+board = esp32doit-devkit-v1
+framework = arduino
+lib_deps = 
+	fastled/FastLED@^3.6.0
+	knolleary/PubSubClient@^2.8
+	ayushsharma82/AsyncElegantOTA@^2.2.8
+	esphome/ESPAsyncWebServer-esphome@^3.1.0
